requirements.txt
ag-ui-protocol==0.1.8
aiohappyeyeballs==2.6.1
aiohttp==3.12.14
aiosignal==1.4.0
annotated-types==0.7.0
anthropic==0.58.2
anyio==4.9.0
argcomplete==3.6.2
asyncpg==0.30.0
attrs==25.3.0
Authlib==1.6.1
beautifulsoup4==4.13.4
boto3==1.39.10
botocore==1.39.10
cachetools==5.5.2
certifi==2025.7.14
cffi==1.17.1
charset-normalizer==3.4.2
click==8.2.1
cohere==5.16.1
colorama==0.4.6
cryptography==45.0.5
cyclopts==3.22.5
ddgs==9.5.4
distro==1.9.0
dnspython==2.7.0
docstring_parser==0.17.0
docutils==0.22
email_validator==2.2.0
eval_type_backport==0.2.2
exceptiongroup==1.3.0
fastapi==0.116.1
fastavro==1.11.1
fastmcp==2.11.0
filelock==3.18.0
filetype==1.2.0
frozenlist==1.7.0
fsspec==2025.7.0
google-auth==2.40.3
google-genai==1.26.0
griffe==1.7.3
groq==0.30.0
h11==0.16.0
hf-xet==1.1.5
httpcore==1.0.9
httpx==0.28.1
httpx-sse==0.4.0
huggingface-hub==0.33.4
idna==3.10
importlib_metadata==8.7.0
isodate==0.7.2
jiter==0.10.0
jmespath==1.0.1
Js2Py_3.13==0.74.1
jsonschema==4.25.0
jsonschema-path==0.3.4
jsonschema-specifications==2025.4.1
lazy-object-proxy==1.11.0
logfire-api==3.25.0
lxml==6.0.0
m3u8==6.0.0
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mcp==1.12.0
mdurl==0.1.2
mistralai==1.9.2
more-itertools==10.7.0
multidict==6.6.3
oauthlib==3.3.1
openai==1.97.1
openapi-core==0.19.5
openapi-pydantic==0.5.1
openapi-schema-validator==0.6.3
openapi-spec-validator==0.7.2
opentelemetry-api==1.35.0
packaging==25.0
parse==1.20.2
pathable==0.4.4
praw==7.8.1
prawcore==2.4.0
primp==0.15.0
prompt_toolkit==3.0.51
propcache==0.3.2
psycopg2==2.9.10
psycopg2-binary==2.9.10
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
pydantic==2.11.7
pydantic-ai==0.4.4
pydantic-ai-slim==0.4.4
pydantic-evals==0.4.4
pydantic-graph==0.4.4
pydantic-settings==2.10.1
pydantic_core==2.33.2
Pygments==2.19.2
pyjsparser==2.7.1
pyotp==2.9.0
pyperclip==1.9.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-multipart==0.0.20
PyYAML==6.0.2
referencing==0.36.2
requests==2.32.4
requests-oauthlib==2.0.0
rfc3339-validator==0.1.4
rich==14.0.0
rich-rst==1.3.1
rpds-py==0.26.0
rsa==4.9.1
s3transfer==0.13.1
six==1.17.0
sniffio==1.3.1
socksio==1.0.0
soupsieve==2.7
SQLAlchemy==2.0.41
sse-starlette==2.4.1
starlette==0.47.2
tenacity==8.5.0
tokenizers==0.21.2
tqdm==4.67.1
tweepy==4.16.0
twikit==2.3.3
types-requests==2.32.4.20250611
typing-inspection==0.4.1
typing_extensions==4.14.1
tzlocal==5.3.1
update-checker==0.18.0
urllib3==2.5.0
uvicorn==0.35.0
wcwidth==0.2.13
websocket-client==1.8.0
websockets==15.0.1
webvtt-py==0.5.1
Werkzeug==3.1.1
x-twitter-mcp==0.1.11
yarl==1.20.1
zipp==3.23.0